Refactoring code into yaacs codestyle
