+++ internal/frontend/main/pages/courses.go
