+++ internal/frontend/main/pages/data.go
